backend/pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bookpoint-backend"
version = "0.1.0"
description = "BookPoint backend foundation (Phase 1)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.13.0",
  "celery>=5.4.0",
  "fastapi>=0.115.0",
  "passlib[bcrypt]>=1.7.4",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.5.0",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.9",
  "redis>=5.0.0",
  "sqlalchemy>=2.0.0",
  "uvicorn[standard]>=0.30.0",
  "email-validator>=2.2.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-q"
